parameterized saml attrib names, and moved code back into authentication.py
